requirements.txt
Flask
python-dotenv
PyPDF2
requests
werkzeug
google-genai
gunicorn
